pyproject.toml
[tool.poetry]
name = "skoltech-infrastructure"
version = "0.1.0"
description = ""
authors = ["Профилактика"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
pillow = "^10.1.0"
imageio = "^2.33.0"
opencv-python = "^4.8.1.78"
scikit-learn = "^1.3.2"
segmentation-models-pytorch = "^0.3.3"
albumentations = "^1.3.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = {extras = ["d"], version = "^23.11.0"}
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
matplotlib = "^3.8.2"
tqdm = "^4.66.1"
seaborn = "^0.13.0"


[tool.poetry.group.research.dependencies]
torchgeo = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
